chore: Refactoring logging tests
