+++ platform-sdk/swirlds-logging/build.gradle.kts
@@ -26,6 +26,7 @@ mainModuleInfo { annotationProcessor("com.google.auto.service.processor") }
 testModuleInfo {
     requires("org.apache.logging.log4j.core")
     requires("com.swirlds.config.extensions.test.fixtures")
+    requires("com.swirlds.logging.test.fixtures")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
